Refactor using mic demo interface
